backend/pyproject.toml
[project]
name = "backend"
version = "0.1.0"
description = "Add your description here"
readme = "README.md"
requires-python = ">=3.13"
dependencies = [
    "asyncpg>=0.30.0",
    "fastapi[standard]>=0.115.12",
    "loguru>=0.7.3",
    "passlib[bcrypt]>=1.7.4",
    "pydantic-settings>=2.9.1",
    "python-jose>=3.5.0",
]
